+++ src/Resources/MSGraph.Autorest/internal/Get-AzADGroup.ps1
@@ -158,8 +158,8 @@ begin {
         }
         $parameterSet = $PSCmdlet.ParameterSetName
         $mapping = @{
-            Get = 'Az.Resources.MSGraph.private\Get-AzADGroup_Get';
-            List = 'Az.Resources.MSGraph.private\Get-AzADGroup_List';
+            Get = 'Az.MSGraph.private\Get-AzADGroup_Get';
+            List = 'Az.MSGraph.private\Get-AzADGroup_List';
         }
 
         $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand(($mapping[$parameterSet]), [System.Management.Automation.CommandTypes]::Cmdlet)