+++ src/Compute/Compute/ChangeLog.md
@@ -20,8 +20,8 @@
 
 -->
 ## Upcoming Release
+* Updated `Set-AzVmExtension` cmdlet to properly display `-Name` and `-Location` parameters as mandatory.
 * Edited `New-AzVmssConfig` second example so it runs successfully by changing the Tag input to the correct format. 
-* Update `Set-AzVmExtension` cmdlet to properly display `-Name` and `-Location` parameters as mandatory.
 
 ## Version 4.22.0
 * Updated `UserData` parameter in VM and VMSS cmdlets to pipe by the Property Name to ensure piping scenarios occur correctly.