+++ src/Compute/Compute/ChangeLog.md
@@ -20,7 +20,14 @@
 
 -->
 ## Upcoming Release
-* Changed New-AzVm cmdlet when using the SimpleParameterSet to not create a PublicIPAddress when a PUblicIPAddress name is not provided. 
+* Edited `New-AzVmssConfig` second example so it runs successfully by changing the Tag input to the correct format. 
+
+## Version 4.22.0
+* Updated `UserData` parameter in VM and VMSS cmdlets to pipe by the Property Name to ensure piping scenarios occur correctly.
+* Changed `New-AzVM` cmdlet when using the SimpleParameterSet to not create a `PublicIPAddress` when a `PublicIPAddress` name is not provided.
+* Added `PlatformFaultDomain` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
+* Added `-Feature` parameter for `New-AzGalleryImageDefinition`
+* Added `DiffDiskPlacement` string parameter to `Set-AzVmOSDisk` and `Set-AzVmssStorageProfile` cmdlets.
 
 ## Version 4.21.0
 * Contains updates to the following powershell cmdlets